Move ERC20 tests to ERC20 custom folder (#772)
