+++ test/token/ERC20/PausableToken.test.js
@@ -1,4 +1,4 @@
-import assertRevert from '../helpers/assertRevert';
+import assertRevert from '../../helpers/assertRevert';
 const PausableToken = artifacts.require('PausableTokenMock');
 
 contract('PausableToken', function ([_, owner, recipient, anotherAccount]) {
